// src/enums.ts
// export enum JarvisMsgEn {
//     dispatch_success = '调度节点分配成功',
//     dispatch_fetch_err = '调度节点分配失败',
//     dispatch_code_err = '调度节点分配失败',
//     dispatch_parse_err = '调度节点分配失败',
//     dispatch_empty_err = '调度节点未分配',
//     dispatch_img2img_err = 'img2img失败',
//     dispatch_txt2img_err = 'txt2img失败',
//     dispatch_chat_err = 'chat失败',
//     dispatch_chat_stream_err = 'chat_stream失败',
// }

export enum JarvisMsgEnglish {
    dispatch_success = 'Successfully assigned scheduling nodes',
    dispatch_fetch_err = 'Scheduling node allocation failed',
    dispatch_code_err = 'Scheduling node allocation failed',
    dispatch_parse_err = 'Scheduling node allocation failed',
    dispatch_empty_err = 'Dispatch node not assigned',
    dispatch_img2img_err = 'img2img failed',
    dispatch_txt2img_err = 'txt2img failed',
    dispatch_chat_err = 'chat failed',
    dispatch_chat_stream_err = 'chat_stream failed',
}

export enum JarvisMsg {
    dispatch_success = 'Successfully assigned scheduling nodes',
    dispatch_fetch_err = 'Scheduling node allocation failed',
    dispatch_code_err = 'Scheduling node allocation failed',
    dispatch_parse_err = 'Scheduling node allocation failed',
    dispatch_empty_err = 'Dispatch node not assigned',
    dispatch_img2img_err = 'img2img failed',
    dispatch_txt2img_err = 'txt2img failed',
    dispatch_chat_err = 'chat failed',
    dispatch_chat_stream_err = 'chat_stream failed',
}

export enum JarvisApiUri {
    uri_check = '/jarvis/v1/token/api/check',
}